rename core entity package
(cherry picked from commit 12aeaaa6da4328604f70c7dac7263434fb0fb3ec)

+++ hodor-admin/src/main/java/org/dromara/hodor/admin/service/impl/JobGroupServiceImpl.java
@@ -22,12 +22,16 @@ import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.dromara.hodor.admin.domain.UserInfo;
+import org.dromara.hodor.admin.core.MsgCode;
+import org.dromara.hodor.admin.dto.user.UserInfo;
+import org.dromara.hodor.admin.exception.ServiceException;
 import org.dromara.hodor.admin.service.JobGroupService;
 import org.dromara.hodor.common.utils.DateUtils;
 import org.dromara.hodor.core.PageInfo;
 import org.dromara.hodor.core.entity.JobGroup;
+import org.dromara.hodor.core.entity.JobInfo;
 import org.dromara.hodor.core.mapper.JobGroupMapper;
+import org.dromara.hodor.core.mapper.JobInfoMapper;
 import org.springframework.stereotype.Service;
 
 /**
@@ -42,6 +46,8 @@ import org.springframework.stereotype.Service;
 public class JobGroupServiceImpl implements JobGroupService {
 
     private final JobGroupMapper jobGroupMapper;
+
+    private final JobInfoMapper jobInfoMapper;
 
     @Override
     public PageInfo<JobGroup> queryGroupListPaging(UserInfo user, String groupName, Integer pageNo, Integer pageSize) {
@@ -72,7 +78,13 @@ public class JobGroupServiceImpl implements JobGroupService {
     }
 
     @Override
-    public void deleteJobGroup(UserInfo user, int id) {
+    public void deleteJobGroup(UserInfo user, Long id) {
+        final JobGroup jobGroup = jobGroupMapper.selectById(id);
+        final Long count = jobInfoMapper.selectCount(Wrappers.<JobInfo>lambdaQuery()
+            .eq(JobInfo::getGroupName, jobGroup.getGroupName()));
+        if (count > 0) {
+            throw new ServiceException(MsgCode.DELETE_GROUP_ERROR, "Current group has {} jobs", count);
+        }
         jobGroupMapper.deleteById(id);
     }
